+++ modules/AWSPowerShell/Cmdlets/SsmSap/Basic/Get-SMSAPResourcePermission-Cmdlet.cs
@@ -54,7 +54,7 @@ namespace Amazon.PowerShell.Cmdlets.SMSAP
         #region Parameter ResourceArn
         /// <summary>
         /// <para>
-        /// The service has not provided documentation for this parameter; please refer to the service's API reference documentation for the latest available information.
+        /// <para>The Amazon Resource Name (ARN) of the resource.</para>
         /// </para>
         /// </summary>
         #if !MODULAR