+++ modules/AWSPowerShell/Cmdlets/SsmSap/Basic/Unregister-SMSAPApplication-Cmdlet.cs
@@ -44,7 +44,7 @@ namespace Amazon.PowerShell.Cmdlets.SMSAP
         #region Parameter ApplicationId
         /// <summary>
         /// <para>
-        /// The service has not provided documentation for this parameter; please refer to the service's API reference documentation for the latest available information.
+        /// <para>The ID of the application.</para>
         /// </para>
         /// </summary>
         #if !MODULAR